HA logic

+++ main/idle.c
@@ -7,9 +7,9 @@ appidle (app_t * a)
 {
    uint8_t l = a->bright;
    if (a->stop)
-      l = a->bright * a->stop / a->fade;
+      l = (int)a->bright * a->stop / a->fade;
    else if (a->fade && a->cycle < a->fade)
-      l = a->bright * a->cycle / a->fade;
+      l = (int)a->bright * a->cycle / a->fade;
    for (int i = 0; i < a->len; i++)
       setl (a->start + i, a, l);
    return NULL;
